Poprawienie środowiska symulacji o pracę z ciągłymi rozkładami prawdopodobieństw.
Potwierdzono zgodność uzyskiwanych czasów i długości kolejek
dla czasów teoretycznych.

+++ src/pl/wroc/pwr/iis/polling/model/sterowanie/FunkcjaOceny_I.java
@@ -13,5 +13,5 @@ public interface FunkcjaOceny_I {
 	 * @param s
 	 * @return
 	 */
-	float ocenaSytuacji(Serwer serwer);
+	double ocenaSytuacji(Serwer serwer);
 }
