+++ src/pl/wroc/pwr/iis/polling/model/sterowanie/deterministyczne/SterownikLosowy.java
@@ -16,7 +16,7 @@ public class SterownikLosowy extends Sterownik {
 	
 	
 	@Override
-	public int getDecyzjaSterujaca(float ocenaStanu, int[] stan, int iloscAkcji) {
+	public int getDecyzjaSterujaca(double ocenaStanu, int[] stan, int iloscAkcji) {
 		int akcja = (int)(RandomGenerator.getDefault().nextInt(iloscAkcji));
 		return akcja;
 	}
