+++ src/pl/wroc/pwr/iis/polling/model/sterowanie/deterministyczne/SterownikNaZmiane.java
@@ -11,7 +11,7 @@ public class SterownikNaZmiane extends Sterownik {
 	
 	
 	@Override
-	public int getDecyzjaSterujaca(float ocenaStanu, int[] stan, int iloscAkcji) {
+	public int getDecyzjaSterujaca(double ocenaStanu, int[] stan, int iloscAkcji) {
 		int akcja = (ostatniaAkcja++) % iloscAkcji;
 		return akcja;
 	}
